add fitur scheduling sisi dosen > Slicing ui scheduling dosen > slicing ui kalender dosen > logic crud scheduling dosen

+++ app/src/main/java/com/acaris/features/documents_mahasiswa/domain/usecase/DeleteDocumentUseCase.kt
@@ -7,6 +7,10 @@ class DeleteDocumentUseCase @Inject constructor(
     private val repository: DocumentRepository
 ) {
     suspend operator fun invoke(documentId: String): Result<Boolean> {
+        if (documentId.isBlank()) {
+            return Result.failure(Exception("ID Dokumen tidak valid atau kosong."))
+        }
+
         return repository.deleteDocument(documentId)
     }
 }
